backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0

# ML
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
joblib==1.4.2

# Database
supabase==2.10.0

# Web Scraping
selenium==4.27.1

# Scheduling
APScheduler==3.10.4

# AI Explanations (optional)
google-generativeai==0.8.3

# Dev / testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
python-dotenv==1.0.1
